refactor(html): parser state
