+++ html/parser/state/stack_of_open_elements.rs
@@ -2,82 +2,24 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use core::ops;
+use std::ops;
 
 use dom::node::Node;
 use html_elements::tag_names;
-use infra::{namespace::Namespace, structure::tree::TreeNode};
+use infra::structure::tree::TreeNode;
 
 // ----------- //
 // Énumération //
 // ----------- //
 
-/// 13.2.4.1 The insertion mode
-///
-/// Le mode d'insertion est une variable d'état qui contrôle l'opération
-/// primaire de l'étape de construction de l'arbre.  Le mode d'insertion
-/// affecte la manière dont les tokens sont traités et si les sections
-/// CDATA sont supportées.
-#[derive(Debug)]
-#[derive(Copy, Clone)]
-#[derive(PartialEq)]
-pub(crate) enum InsertionMode {
-    Initial,
-    BeforeHTML,
-    BeforeHead,
-    InHead,
-    InHeadNoscript,
-    AfterHead,
-    InBody,
-    Text,
-    InTable,
-    InTableText,
-    InCaption,
-    InColumnGroup,
-    InTableBody,
-    InRow,
-    InCell,
-    InSelect,
-    InSelectInTable,
-    InTemplate,
-    AfterBody,
-    InFrameset,
-    AfterFrameset,
-    AfterAfterBody,
-    AfterAfterFrameset,
-}
-
-/// 13.2.4.2 The stack of open elements
 #[derive(Default)]
 pub(crate) struct StackOfOpenElements {
     elements: Vec<TreeNode<Node>>,
 }
 
-/// 13.2.4.3 The list of active formatting elements
-#[derive(Default)]
-pub(crate) struct ListOfActiveFormattingElements {
-    entries: Vec<Entry>,
-}
-
-// ----------- //
-// Énumération //
-// ----------- //
-
-#[derive(PartialEq)]
-pub(crate) enum Entry {
-    Marker,
-    Element(TreeNode<Node>),
-}
-
 // -------------- //
 // Implémentation //
 // -------------- //
-
-impl InsertionMode {
-    pub(crate) fn switch_to(&mut self, mode: Self) {
-        *self = mode;
-    }
-}
 
 impl StackOfOpenElements {
     ///   - applet
@@ -257,25 +199,6 @@ impl StackOfOpenElements {
         [tag_names::html, tag_names::table, tag_names::template]
     }
 
-    pub(crate) fn topmost_special_node_below(
-        &self,
-        formatting_element: &TreeNode<Node>,
-        is_special_tag: impl Fn(tag_names, Namespace) -> bool,
-    ) -> Option<(usize, &TreeNode<Node>)> {
-        self.elements.iter().enumerate().rfind(|&(_, node)| {
-            if node == formatting_element {
-                return false;
-            }
-
-            is_special_tag(
-                node.element_ref().tag_name(),
-                node.element_ref()
-                    .namespace()
-                    .expect("Devrait être un nom de namespace valide"),
-            )
-        })
-    }
-
     /// <https://github.com/rust-lang/rust/issues/83701>
     fn scoped_elements_with<const N: usize>(
         list: impl IntoIterator<Item = tag_names>,
@@ -297,109 +220,9 @@ impl StackOfOpenElements {
     }
 }
 
-impl ListOfActiveFormattingElements {
-    pub(crate) fn clear_up_to_the_last_marker(&mut self) {
-        while let Some(entry) = self.entries.pop() {
-            if entry.is_marker() {
-                break;
-            }
-        }
-    }
-
-    pub(crate) fn contains_element(
-        &self,
-        element: &TreeNode<Node>,
-    ) -> bool {
-        self.entries
-            .iter()
-            .any(|entry| Entry::Element(element.to_owned()).eq(entry))
-    }
-
-    pub(crate) fn insert_marker_at_end(&mut self) {
-        self.entries.push(Entry::Marker);
-    }
-
-    pub(crate) fn last_element_before_marker(
-        &self,
-        tag_name: tag_names,
-    ) -> Option<(usize, TreeNode<Node>)> {
-        self.entries
-            .iter()
-            .enumerate()
-            .rfind(|(_, entry)| {
-                if let Entry::Element(element) = entry {
-                    tag_name == element.element_ref().local_name()
-                } else {
-                    false
-                }
-            })
-            .and_then(|(idx, entry)| {
-                entry.element().map(|node| (idx, node.to_owned()))
-            })
-    }
-
-    pub(crate) fn remove_element(&mut self, element: &TreeNode<Node>) {
-        if let Some(idx) = self.entries.iter().position(|entry| {
-            if let Entry::Element(element_node) = entry {
-                element_node == element
-            } else {
-                false
-            }
-        }) {
-            self.entries.remove(idx);
-        }
-    }
-
-    pub(crate) fn position_of(
-        &self,
-        element: &TreeNode<Node>,
-    ) -> Option<usize> {
-        self.entries.iter().enumerate().rposition(|(_, entry)| {
-            if let Entry::Element(element_ref) = entry {
-                element_ref == element
-            } else {
-                false
-            }
-        })
-    }
-}
-
-impl Entry {
-    pub(crate) const fn is_element(&self) -> bool {
-        matches!(self, Self::Element(_))
-    }
-
-    pub(crate) const fn is_marker(&self) -> bool {
-        matches!(self, Self::Marker)
-    }
-
-    pub(crate) const fn element(&self) -> Option<&TreeNode<Node>> {
-        match self {
-            | Entry::Marker => None,
-            | Entry::Element(node) => Some(node),
-        }
-    }
-
-    pub(crate) const fn element_unchecked(&self) -> &TreeNode<Node> {
-        match self {
-            | Entry::Marker => {
-                panic!("N'est pas une entrée de type Entry::Element.")
-            }
-            | Entry::Element(node) => node,
-        }
-    }
-}
-
 // -------------- //
 // Implémentation // -> Interface
 // -------------- //
-
-impl Default for InsertionMode {
-    /// Initialement, le mode d'insertion est "initial".
-    fn default() -> Self {
-        Self::Initial
-    }
-}
 
 impl ops::Deref for StackOfOpenElements {
     type Target = Vec<TreeNode<Node>>;
@@ -412,19 +235,5 @@ impl ops::Deref for StackOfOpenElements {
 impl ops::DerefMut for StackOfOpenElements {
     fn deref_mut(&mut self) -> &mut Self::Target {
         self.elements.as_mut()
-    }
-}
-
-impl ops::Deref for ListOfActiveFormattingElements {
-    type Target = Vec<Entry>;
-
-    fn deref(&self) -> &Self::Target {
-        self.entries.as_ref()
-    }
-}
-
-impl ops::DerefMut for ListOfActiveFormattingElements {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.entries.as_mut()
     }
 }
